refactor: upgrade hybridly config
