+++ .npmrc
@@ -1,1 +1,3 @@
 shamefully-hoist=true
+node-linker=hoisted
+use-lockfile-v6=true
