+++ config/typescript-transformer.php
@@ -48,7 +48,7 @@ return [
      * The package will write the generated TypeScript to this file.
      */
 
-    'output_file' => resource_path('types/data-transfer-objects.d.ts'),
+    'output_file' => base_path('.hybridly/back-end.d.ts'),
 
     /*
      * When the package is writing types to the output file, a writer is used to
