+++ package.json
@@ -6,44 +6,40 @@
   },
   "devDependencies": {
     "@fontsource/poppins": "^4.5.10",
-    "@formkit/auto-animate": "1.0.0-beta.5",
-    "@iconify-json/ant-design": "^1.1.4",
-    "@iconify-json/fa6-solid": "^1.1.10",
-    "@iconify-json/fluent": "^1.1.17",
-    "@iconify-json/gg": "^1.1.3",
-    "@iconify-json/heroicons-solid": "^1.1.6",
-    "@iconify-json/ic": "^1.1.12",
-    "@iconify-json/icon-park-outline": "^1.1.9",
-    "@iconify-json/ion": "^1.1.6",
-    "@iconify-json/material-symbols": "^1.1.26",
-    "@iconify-json/mdi": "^1.1.41",
-    "@iconify-json/ph": "^1.1.3",
-    "@innocenzi/eslint-config": "^0.14.0",
-    "@intlify/unplugin-vue-i18n": "^0.8.1",
+    "@formkit/auto-animate": "1.0.0-beta.6",
+    "@iconify-json/ant-design": "^1.1.5",
+    "@iconify-json/fa6-solid": "^1.1.12",
+    "@iconify-json/fluent": "^1.1.18",
+    "@iconify-json/gg": "^1.1.4",
+    "@iconify-json/heroicons-solid": "^1.1.7",
+    "@iconify-json/ic": "^1.1.13",
+    "@iconify-json/icon-park-outline": "^1.1.10",
+    "@iconify-json/ion": "^1.1.8",
+    "@iconify-json/material-symbols": "^1.1.31",
+    "@iconify-json/mdi": "^1.1.47",
+    "@iconify-json/ph": "^1.1.4",
+    "@innocenzi/eslint-config": "^0.14.1",
+    "@intlify/unplugin-vue-i18n": "^0.8.2",
     "@tailwindcss/forms": "^0.5.3",
-    "@types/node": "^18.11.18",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "@vue/runtime-core": "^3.2.45",
-    "@vueuse/core": "^9.9.0",
-    "@vueuse/head": "^1.0.22",
+    "@types/node": "^18.14.0",
+    "@vue/runtime-core": "^3.2.47",
+    "@vueuse/core": "^9.13.0",
+    "@vueuse/head": "^1.0.26",
     "autoprefixer": "^10.4.13",
-    "eslint": "^8.31.0",
+    "eslint": "^8.34.0",
     "hybridly": "link:/Users/enzoinnocenzi/Code/projects/hybridly/hybridly/packages/hybridly",
-    "laravel-vite-plugin": "^0.7.3",
-    "postcss": "^8.4.20",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.4",
-    "unplugin-auto-import": "^0.12.1",
-    "unplugin-icons": "^0.14.15",
-    "unplugin-vue-components": "^0.22.12",
-    "vite": "^4.0.3",
-    "vite-plugin-run": "^0.2.1",
-    "vue": "^3.2.45",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.2.7",
+    "typescript": "^4.9.5",
+    "vite": "^4.1.2",
+    "vite-plugin-inspect": "^0.7.15",
+    "vite-plugin-run": "^0.3.2",
+    "vue": "^3.2.47",
     "vue-i18n": "^9.2.2"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.7",
-    "axios": "^1.2.2"
+    "@headlessui/vue": "^1.7.10",
+    "axios": "^1.3.3"
   },
   "postcss": {
     "plugins": {
