+++ resources/application/root.blade.php
