+++ resources/components/like-button.vue
