+++ resources/components/verified-badge.vue
