+++ resources/pages/chirps/show.vue
