--- resources/types/hybridly.d.ts
@@ -1,8 +1,0 @@
-interface GlobalHybridlyProperties extends App.Data.Hybridly.SharedData {
-	flash: {
-		success?: string
-		error?: string
-		warning?: string
-		info?: string
-	}
-}
