+++ tsconfig.json
@@ -1,38 +1,3 @@
 {
-    "compilerOptions": {
-        "target": "esnext",
-        "module": "esnext",
-        "moduleResolution": "node",
-        "strict": true,
-        "jsx": "preserve",
-        "sourceMap": true,
-        "resolveJsonModule": true,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true,
-        "lib": [
-            "esnext",
-            "dom"
-        ],
-        "types": [
-            "vite/client",
-            "unplugin-icons/types/vue"
-        ],
-        "baseUrl": ".",
-        "paths": {
-            "@/*": [
-                "resources/*"
-            ]
-        }
-    },
-    "vueCompilerOptions": {
-        "experimentalSuppressInvalidJsxElementTypeErrors": true
-    },
-    "include": [
-        "resources/**/*"
-    ],
-    "exclude": [
-        "public/**/*",
-        "node_modules",
-        "vendor"
-    ]
+	"extends": "./.hybridly/tsconfig.json"
 }
